+++ src/components/Tournament/Tournament.css
