+++ src/components/TournamentInfo/TournamentInfo.adaptive.css
